Move host pipe compile time properties processing to opt
