+++ llvm/lib/SYCLLowerIR/HostPipes.cpp
@@ -8,8 +8,8 @@
 // See comments in the header.
 //===----------------------------------------------------------------------===//
 
-#include "HostPipes.h"
-#include "CompileTimePropertiesPass.h"
+#include "llvm/SYCLLowerIR/HostPipes.h"
+#include "llvm/SYCLLowerIR/CompileTimePropertiesPass.h"
 
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/StringRef.h"
